Olaris Server
https://gitlab.com/olaris/olaris-docker

Olaris is an open-source, community driven, media manager and transcoding server. The main interface is the olaris-react project although in due time we hope to support multiple clients / applications.

Focus on one feature at a time
We will work on features until they are perfect (or as close to it as possible). We would rather have a product where three features work really well than a product with 30 unfinished features.
This does not mean we won't work in parallel, it simply means we will not start anything new until we are happy the new feature works to a high standard.

Our users are not our product
We don't want to collect metadata, we don't want to sell metadata your data is yours and yours alone.

Singular Focus: Video.
Our sole focus is on video and video alone, anything that does not meet this requirement will not be considered. This means for example we will never add music support due to different approach that would be required throughout the application.